rename root store to app store
