+++ src/app/app-store/app.state.ts
@@ -3,7 +3,7 @@ import { BeneficiariesState } from '../pages/admin/beneficiaries/beneficiaries.s
 import { VolunteersState } from '../pages/admin/volunteers/volunteers.state';
 import { RequestsState } from '../pages/admin/requests/requests.state';
 
-export interface RootState {
+export interface AppState {
   auth: AuthState;
   beneficiaries: BeneficiariesState;
   volunteers: VolunteersState;
