+++ src/app/app.module.ts
@@ -5,7 +5,7 @@ import { appRoutes } from './app.routes';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { RootStoreModule } from '@store/root-store.module';
+import { AppStoreModule } from '@store/app-store.module';
 import {
   HTTP_INTERCEPTORS,
   HttpClient,
@@ -33,7 +33,7 @@ export function HttpLoaderFactory(httpClient: HttpClient) {
     HttpClientModule,
     BrowserAnimationsModule,
     FlexLayoutModule,
-    RootStoreModule,
+    AppStoreModule,
     MatSnackBarModule,
     MatIconModule,
     RouterModule.forRoot(appRoutes),
